fixed task 1 by changing int to float

+++ src/rasterizer.cpp
@@ -76,8 +76,8 @@ namespace CGL {
       float miny = min(y0, min(y1, y2));
       float maxy = max(y0, max(y1, y2));
 
-      for (int x = minx+0.5; x < maxx; x++) {
-          for (int y = miny+0.5; y < maxy; y++) {
+      for (float x = minx+0.5; x < maxx; x++) {
+          for (float y = miny+0.5; y < maxy; y++) {
               float flag1 = (-(x - x0) * (y1 - y0) + (y - y0) * (x1 - x0));
               float flag2 = (-(x - x1) * (y2 - y1) + (y - y1) * (x2 - x1));
               float flag3 = (-(x - x2) * (y0 - y2) + (y - y2) * (x0 - x2));
